requirements.txt
suds==0.4
wsgiref==0.1.2
pandas>=0.15.2
PyYAML>=3.11
xlwt==0.7.5